Added new file structure and designated VolunteerOver as the new default view of the volunteer componenet
