--- src/components/Pages/Volunteers/Overview/index.js
@@ -1,2 +1,0 @@
-import Overview from "./Overview";
-export default Overview;
