+++ src/components/VolunteerComponents/Pages/Dashboard/VolunteerOverview/index.js
@@ -1,0 +1,2 @@
+import Overview from './Overview';
+export default Overview;
