--- src/components/VolunteerComponents/Pages/Dashboard/index.js
@@ -1,2 +1,0 @@
-import VolunteerDashboard from './VolunteerDashboard';
-export default VolunteerDashboard;
